+++ native/font-loader.js
@@ -1,5 +1,4 @@
 /* eslint-env browser */
-/* eslint-disable class-methods-use-this */
 
 // ref: https://gist.github.com/vincentriemer/b8e5e97a92a44638a118016801d378cd