+++ src/ExponentSpeech.dom.js
@@ -1,3 +1,4 @@
+// FIXME "web" module doesn't work for "dom"
 import ExponentSpeech from '../../expo/build/Speech/ExponentSpeech.web';
 
 export default ExponentSpeech;